backend/pyproject.toml
[tool.black]
line-length = 100

[tool.poetry]
name = "brainshare-pelican-backend"
version = "0.1.0"
description = ""
authors = ["Zachary King <zaking17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
celery = { extras = ["redis"], version = "^5.2.7" }
fastapi = "^0.95.2"
glom = "^23.3.0"
sqlmodel = "^0.0.8"
sqlalchemy = { version = "^1.4.41" }
uvicorn = { extras = ["standard"], version = "^0.22.0" }
pyjwt = "^2.7.0"
supabase = "^1.0.3"
storage3 = "^0.5.2"
# sticking w psycopg2 in case we want to use it as the celery backend
psycopg2-binary = "^2.9.6"

[tool.poetry.group.testing.dependencies]
pytest = "^7.2.2"
pytest-asyncio = "^0.20.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"
black = "^23.1.0"
types-redis = "^4.5.4.1"
sqlalchemy-stubs = "^0.4"
sqlacodegen = { git = "https://github.com/agronholm/sqlacodegen", branch = "master" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
